Move SelectedCities to its own folder
