+++ src/components/pages/SelectedCities/index.js
@@ -1,0 +1,1 @@
+export { default as SelectedCities } from "./SelectedCitiesContainer";
